+++ src/main/java/goorm_runner/backend/member/application/AuthorityRepository.java
@@ -1,5 +1,7 @@
-package goorm_runner.backend.member.domain;
+package goorm_runner.backend.member.application;
 
+import goorm_runner.backend.member.domain.Authority;
+import goorm_runner.backend.member.domain.AuthorityType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
