+++ src/main/java/goorm_runner/backend/member/application/MemberAuthorityRepository.java
@@ -1,5 +1,6 @@
-package goorm_runner.backend.member.domain;
+package goorm_runner.backend.member.application;
 
+import goorm_runner.backend.member.domain.MemberAuthority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberAuthorityRepository extends JpaRepository<MemberAuthority, Long> {
