+++ src/main/java/goorm_runner/backend/member/application/MemberRepository.java
@@ -1,5 +1,6 @@
-package goorm_runner.backend.member.domain;
+package goorm_runner.backend.member.application;
 
+import goorm_runner.backend.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
